+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Changelog
 ## Unreleased
+### added
+* record 可以直接处理内存中的 DataFrame 账单。
+
 ### fixed
 * evaluate 类的可视化函数补充 rendered 选项
 * 修正了 tradevolume 周数和 pandas 给出的 iso 周数不一定对应的问题
+* 修复了超高价转债债券收益率求解的 runtimeerror，直接返回 None
 
 ## v0.9.3 - 2020.06.20
 ### added